Rename header files to CamelCase
