+++ standard/Platform/Common.h
